+++ SerializeReferenceEditor/Assets/SREditor/Package/Editor/Scripts/DoubleCleaner/SRDuplicateCleaner.cs
@@ -53,12 +53,15 @@ namespace SerializeReferenceEditor.Editor.DoubleCleaner
 				if (property.propertyType == SerializedPropertyType.ManagedReference)
 				{
 					var managedReferenceValue = property.managedReferenceValue;
+					
 					if (managedReferenceValue != null && !seenObjects.Add(managedReferenceValue))
 					{
+						var refChanged = false;
 						switch (duplicateMode)
 						{
 							case SRDuplicateMode.Null:
 								property.managedReferenceValue = null;
+								refChanged = true;
 								break;
 
 							case SRDuplicateMode.Default:
@@ -66,6 +69,7 @@ namespace SerializeReferenceEditor.Editor.DoubleCleaner
 								if (currentValue == null)
 								{
 									property.managedReferenceValue = null;
+									refChanged = true;
 									break;
 								}
 
@@ -106,20 +110,24 @@ namespace SerializeReferenceEditor.Editor.DoubleCleaner
 											if (field != null)
 											{
 												property.managedReferenceValue = GetDefaultValueFromField(field);
+												refChanged = true;
 											}
 											else
 											{
 												property.managedReferenceValue = null;
+												refChanged = true;
 											}
 										}
 										else
 										{
 											property.managedReferenceValue = null;
+											refChanged = true;
 										}
 									}
 									else
 									{
 										property.managedReferenceValue = null;
+										refChanged = true;
 									}
 								}
 								else
@@ -133,6 +141,7 @@ namespace SerializeReferenceEditor.Editor.DoubleCleaner
 									}
 									
 									property.managedReferenceValue = newInstance;
+									refChanged = true;
 								}
 								break;
 
@@ -143,16 +152,19 @@ namespace SerializeReferenceEditor.Editor.DoubleCleaner
 									var newInstance = CreateDeepCopy(managedReferenceValue);
 									property.managedReferenceValue = newInstance;
 									seenObjects.Add(newInstance);
+									refChanged = true;
 								}
 								catch (Exception e)
 								{
 									Debug.LogError($"Failed to create instance of type {sourceType.Name}: {e.Message}");
 									property.managedReferenceValue = null;
+									refChanged = true;
 								}
 								break;
 						}
 						
-						property.serializedObject.ApplyModifiedProperties();
+						if (refChanged)
+							property.serializedObject.ApplyModifiedProperties();
 					}
 				}
 			}