+++ SerializeReferenceEditor/Assets/SREditor/Package/Scripts/SRHiddenAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace SerializeReferenceEditor
+{
+	[AttributeUsage(AttributeTargets.Class)]
+	public class SRHiddenAttribute : Attribute
+	{
+	}
+}