+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/ComplexData.cs
@@ -1,9 +1,10 @@
+using System;
 using SerializeReferenceEditor;
 using UnityEngine;
 
 namespace Demo
 {
-	[SRName("Data/Custom types/Complex")]
+	[Serializable, SRName("Data/Custom types/Complex")]
 	public class ComplexData : AbstractData
 	{
 		[SerializeReference, SR]