+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/FloatData.cs
@@ -1,9 +1,10 @@
-﻿using SerializeReferenceEditor;
+﻿using System;
+using SerializeReferenceEditor;
 using UnityEngine;
 
 namespace Demo
 {
-	[SRName("Data/Simple types/Float")]
+	[Serializable, SRName("Data/Simple types/Float")]
 	public class FloatData : AbstractData
 	{
 		[Range(0f, 1f)]