+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/FloatNamedData.cs
@@ -1,9 +1,10 @@
+using System;
 using SerializeReferenceEditor;
 using UnityEngine;
 
 namespace Demo
 {
-	[SRName("Data/Simple types/Named Float")]
+	[Serializable, SRName("Data/Simple types/Named Float")]
 	public class FloatNamedData : AbstractNamedData
 	{
 		[Range(0f, 1f)]