+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/HiddenData.cs
@@ -1,0 +1,11 @@
+using System;
+using SerializeReferenceEditor;
+
+namespace Demo
+{
+	[Serializable, SRName("Data/Simple types/Hidden"), SRHidden]
+	public class HiddenData : AbstractData
+	{
+		public int YouCantShowMe;
+	}
+}