+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/IntegerNamedData.cs
@@ -1,8 +1,9 @@
+using System;
 using SerializeReferenceEditor;
 
 namespace Demo
 {
-	[SRName("Data/Simple types/Named Integer")]
+	[Serializable, SRName("Data/Simple types/Named Integer")]
 	public class IntegerNamedData : AbstractNamedData
 	{
 		public int Int;