+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/StringData.cs
@@ -1,8 +1,9 @@
-﻿using SerializeReferenceEditor;
+﻿using System;
+using SerializeReferenceEditor;
 
 namespace Demo
 {
-	[SRName("Data/Simple types/String")]
+	[Serializable, SRName("Data/Simple types/String")]
 	public class StringData : AbstractData
 	{
 		public string Str;