+++ SerializeReferenceEditor/Assets/SREditor/Samples/Demo/Datas/StringNamedData.cs
@@ -1,8 +1,9 @@
+using System;
 using SerializeReferenceEditor;
 
 namespace Demo
 {
-	[SRName("Data/Simple types/Named String")]
+	[Serializable, SRName("Data/Simple types/Named String")]
 	public class StringNamedData : AbstractNamedData
 	{
 		public string Value;