+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.3.2
+
+- Added SRHidden attribute, for hide type at select list
+- Update Demo files
+
+# 1.3.1
+
+- Added SRDrawerOptions to Draw method
+
 # 1.3.0
 
 New Features: